Refactor server - extract handlers
